Ensure unique history for numbered buffers
Prevously the history file was set before the buffer was created so
numbered buffers would all share the same history. Now it's set after
the buffer so it'll always be unique for each buffer. No more lost
history due to buffers overwriting the history from other buffers.

This also means that when using the prefix, instead of asking for the
file it'll ask for the directory which is one less component to
type/worry about.